feat: add runtime profiling, analysis tools, and enhance d-sweep infrastructure
New analysis tools:
- Add run_profile_5d.py for quick Langevin training benchmarks with formatted output
- Add analyze_d_sweep.py for Hessian eigenvalue analysis across d-sweep with QB decomposition
- Add diagnostics_langevin_runtime.py for comprehensive FCN3 runtime profiling with Julia integration

Enhance d-sweep runner (d_sweep_fcn3_P4d.py):
- Switch to log-spaced d values (default 10 points) from linear steps
- Change P scaling from 4*d to 5*d
- Scale learning rate with P (lrA * P)
- Add dual-GPU round-robin job distribution (cuda:0, cuda:1)
- Support running on pre-existing experiment folders via EXPERIMENT_FILES
- Add --experiment_dirname flag for organizing sweep outputs

Improve ensembling script (ensembling_fcn3_erf_cubic.py):
- Switch from float64 to float32 for faster computation
- Reduce logging frequency (10k → 100k epochs for eigenvalues/gsheets)
- Add --timestamped flag to control timestamp in folder names (default: False)
- Add --experiment_dirname flag for custom experiment organization
- Persist all CLI args to config.json in experiment folder
- Add ghost mode infrastructure for runs without saving/logging

+++ legacy/3layerexps/ensembling_fcn3.py
@@ -186,7 +186,7 @@ def delete_gsheets_row_by_foldername(folder_name):
         print(f'Google Sheets row delete error: {e}')
 
 # Set the default dtype to float64
-torch.set_default_dtype(torch.float64)
+torch.set_default_dtype(torch.float32)
 
 # Custom loss function (slightly faster than MSE)
 def custom_mse_loss(outputs, targets):
@@ -252,7 +252,7 @@ if __name__ == '__main__':
     global current_base_learning_rate
     current_base_learning_rate = lrA
     # Set the default dtype to float64
-    torch.set_default_dtype(torch.float64)
+    torch.set_default_dtype(torch.float32)
 
     if not debug:
         if args.modeldesc is not None:
@@ -385,7 +385,7 @@ if __name__ == '__main__':
     # Seed for data
     torch.manual_seed(DATA_SEED)
     # Create data on GPU directly and pin memory
-    X = torch.randn((num_samples, input_size), dtype=torch.float64, device=device)
+    X = torch.randn((num_samples, input_size), dtype=torch.float32, device=device)
     Y = X[:, 0].unsqueeze(-1)
 
     # Seed for model
@@ -410,10 +410,10 @@ if __name__ == '__main__':
     losses = []
     
     # Move weight decay to GPU - THIS IS CRITICAL FOR PERFORMANCE
-    weight_decay = torch.tensor([input_size, hidden_size, hidden_size*chi], dtype=torch.float64, device=device) * t
+    weight_decay = torch.tensor([input_size, hidden_size, hidden_size*chi], dtype=torch.float32, device=device) * t
 
     # Pre-allocate noise buffer for Langevin dynamics
-    noise_buffer = torch.empty(1, device=device, dtype=torch.float64)
+    noise_buffer = torch.empty(1, device=device, dtype=torch.float32)
     
     # Resume logic
     epoch = 0
